Finished Client Queue and Reservation

+++ FoodReserve.API/FoodReserve.AdminPortal/Services/CustomerService.cs
@@ -23,53 +23,53 @@ namespace FoodReserve.AdminPortal.Services
 
         public async Task<PaginatedResponse<IEnumerable<CustomerResponse>>> GetAllAsync(int pageNumber, int pageSize, string? keyword)
         {
-            var response = await _httpClient.GetAsync($"api/customer?pageNumber={pageNumber}&pageSize={pageSize}&keyword={keyword}");
+            var response = await _httpClient.GetAsync($"api/admin/customer?pageNumber={pageNumber}&pageSize={pageSize}&keyword={keyword}");
             response.EnsureSuccessStatusCode();
             return (await response.Content.ReadFromJsonAsync<PaginatedResponse<IEnumerable<CustomerResponse>>>(_jsonOptions))!;
         }
 
         public async Task<PaginatedResponse<IEnumerable<UserResponse>>> GetAllUserAsync(int pageNumber, int pageSize, string? keyword)
         {
-            var response = await _httpClient.GetAsync($"api/customer/user?pageNumber={pageNumber}&pageSize={pageSize}&keyword={keyword}");
+            var response = await _httpClient.GetAsync($"api/admin/customer/user?pageNumber={pageNumber}&pageSize={pageSize}&keyword={keyword}");
             response.EnsureSuccessStatusCode();
             return (await response.Content.ReadFromJsonAsync<PaginatedResponse<IEnumerable<UserResponse>>>(_jsonOptions))!;
         }
 
         public async Task<CustomerResponse> GetByIdAsync(string id)
         {
-            var response = await _httpClient.GetAsync($"api/customer/{id}");
+            var response = await _httpClient.GetAsync($"api/admin/customer/{id}");
             response.EnsureSuccessStatusCode();
             return (await response.Content.ReadFromJsonAsync<CustomerResponse>(_jsonOptions))!;
         }
 
         public async Task<UserResponse> GetUserByIdAsync(string id)
         {
-            var response = await _httpClient.GetAsync($"api/customer/user/{id}");
+            var response = await _httpClient.GetAsync($"api/admin/customer/user/{id}");
             response.EnsureSuccessStatusCode();
             return (await response.Content.ReadFromJsonAsync<UserResponse>(_jsonOptions))!;
         }
 
         public async Task CreateAsync(CustomerRequest customer)
         {
-            var response = await _httpClient.PostAsJsonAsync("api/customer", customer, _jsonOptions);
+            var response = await _httpClient.PostAsJsonAsync("api/admin/customer", customer, _jsonOptions);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task UpdateAsync(string id, CustomerRequest customer)
         {
-            var response = await _httpClient.PutAsJsonAsync($"api/customer/{id}", customer, _jsonOptions);
+            var response = await _httpClient.PutAsJsonAsync($"api/admin/customer/{id}", customer, _jsonOptions);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task ToggleStatusAsync(string id)
         {
-            var response = await _httpClient.PostAsync($"api/customer/toggle/{id}", null);
+            var response = await _httpClient.PostAsync($"api/admin/customer/toggle/{id}", null);
             response.EnsureSuccessStatusCode();
         }
 
         public async Task DeleteAsync(string id)
         {
-            var response = await _httpClient.DeleteAsync($"api/customer/{id}");
+            var response = await _httpClient.DeleteAsync($"api/admin/customer/{id}");
             response.EnsureSuccessStatusCode();
         }
